move defaults into scoring
